+++ test/helpers/JavaScripts.js
@@ -1,6 +1,6 @@
 import {Script} from 'vm';
 
-export function Scripts() {
+export function Scripts(enableDummy = true) {
   const scripts = {};
 
   return {
@@ -9,13 +9,14 @@ export function Scripts() {
     compile,
   };
 
-  function register({id, type, behaviour}) {
+  function register({id, type, behaviour, logger}) {
     let scriptBody, language;
 
     switch (type) {
       case 'bpmn:SequenceFlow': {
         if (!behaviour.conditionExpression) return;
         language = behaviour.conditionExpression.language;
+        if (!language) return;
         scriptBody = behaviour.conditionExpression.body;
         break;
       }
@@ -25,29 +26,50 @@ export function Scripts() {
       }
     }
 
-    const compiled = compile(language, `${type}/${id}`, scriptBody);
-    if (!compiled) return;
-    scripts[id] = compiled;
 
-    return compiled;
+    if (!language || !scriptBody) {
+      if (!enableDummy) return;
+      const script = dummyScript(language, `${type}/${id}`, logger);
+      scripts[id] = script;
+      return script;
+    }
+
+    if (!/^javascript$/i.test(language)) return;
+
+    const script = javaScript(language, `${type}/${id}`, scriptBody);
+    scripts[id] = script;
+
+    return script;
   }
 
   function compile(language, filename, scriptBody) {
-    if (!/^javascript$/i.test(language)) return;
     return new Script(scriptBody, {filename});
   }
 
   function getScript(language, {id}) {
-    if (!/^javascript$/i.test(language)) return;
-    const script = scripts[id];
-    if (!script) return;
+    return scripts[id];
+  }
 
+  function javaScript(language, filename, scriptBody) {
+    const script = new Script(scriptBody, {filename});
     return {
-      execute,
+      script,
+      language,
+      execute(executionContext, callback) {
+        return script.runInNewContext({...executionContext, next: callback});
+      }
     };
+  }
 
-    function execute(executionContext, callback) {
-      return script.runInNewContext({...executionContext, next: callback});
-    }
+  function dummyScript(language) {
+    return {
+      isDummy: true,
+      language,
+      execute(executionContext, callback) {
+        const {id, executionId} = executionContext.content;
+        executionContext.environment.Logger('dummy:script').debug(`<${executionId} (${id})> passthrough dummy script ${language || 'esperanto'}`);
+        callback();
+      },
+    };
   }
 }